requirements.txt
pygame==2.6.1
sounddevice
numpy==1.26.4
torch
torchaudio
torchvision
transformers==4.39.3
librosa==0.9.2
pydub==0.25.1
pyopenjtalk-dict==0.3.4.dev2
pyworld-prebuilt==0.3.5.post1
faster-whisper==0.10.1
ctranslate2==3.24.0
sentencepiece==0.2.0
requests==2.32.3
pillow==11.0.0
pyyaml==6.0.2
tqdm==4.67.1
openai-whisper==20240930
tweepy
gspread
google-auth
google-genai
deep-translator
pytz
